Use the first frame of an interacton to base st/at.
This only matters when a frame is very slow - like if a ton of
rendering has to be done, on an unpredicted image load. If
rendering the first frame took .1 seconds, the second frame of
a transform and move would be .1 later than the first.

Now, it's 1 frame later - the .1s becomes time spent before
the first frame, rather than time spend between frames.